adjusting architecture for better developability and maintenance
