+++ whittle/src/interfaces/case_structure_interface.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+class ICaseStructureManager(ABC):
+    @abstractmethod
+    def setup_case_structure(self) -> None: pass
