+++ whittle/src/interfaces/conversation_interface.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+class IAIConversationManager(ABC):
+    @abstractmethod
+    def get_response(self, user_input: str) -> str: pass
