+++ whittle/src/interfaces/mesh_executor_interface.py
@@ -1,0 +1,5 @@
+from abc import ABC, abstractmethod
+
+class IMeshExecutor(ABC):
+    @abstractmethod
+    def run_mesh(self) -> None: pass
